Fix timing issue
Things didn't work previously because of the timing of the init() function, which was after propchange events had already taken place. This change makes the defineEvents() mixin independent of when its init function runs since it handles both existing values and future changes. The code is a little repetitive but oh well.